Added a feature flag for the in-memory Machine

+++ src/SIL.XForge.Scripture/Services/MachineProjectService.cs
@@ -74,7 +74,12 @@ namespace SIL.XForge.Scripture.Services
                 SourceLanguageTag = project.TranslateConfig.Source.WritingSystem.Tag,
                 TargetLanguageTag = project.WritingSystem.Tag,
             };
-            await _engineService.AddProjectAsync(machineProject);
+
+            // Only add to the in-memory instance if it is enabled
+            if (await _featureManager.IsEnabledAsync(FeatureFlags.MachineInMemory))
+            {
+                await _engineService.AddProjectAsync(machineProject);
+            }
 
             // Ensure that the Machine API feature flag is enabled
             if (!await _featureManager.IsEnabledAsync(FeatureFlags.MachineApi))
@@ -105,8 +110,11 @@ namespace SIL.XForge.Scripture.Services
 
         public async Task BuildProjectAsync(string curUserId, string sfProjectId, CancellationToken cancellationToken)
         {
-            // Build the project with the in process Machine instance
-            await _engineService.StartBuildByProjectIdAsync(sfProjectId);
+            // Build the project with the in-memory Machine instance
+            if (await _featureManager.IsEnabledAsync(FeatureFlags.MachineInMemory))
+            {
+                await _engineService.StartBuildByProjectIdAsync(sfProjectId);
+            }
 
             // Ensure that the Machine API feature flag is enabled
             if (!await _featureManager.IsEnabledAsync(FeatureFlags.MachineApi))
@@ -142,8 +150,11 @@ namespace SIL.XForge.Scripture.Services
 
         public async Task RemoveProjectAsync(string curUserId, string sfProjectId, CancellationToken cancellationToken)
         {
-            // Remove the project from the in process Machine instance
-            await _engineService.RemoveProjectAsync(sfProjectId);
+            // Remove the project from the in-memory Machine instance
+            if (await _featureManager.IsEnabledAsync(FeatureFlags.MachineInMemory))
+            {
+                await _engineService.RemoveProjectAsync(sfProjectId);
+            }
 
             // Ensure that the Machine API feature flag is enabled
             if (!await _featureManager.IsEnabledAsync(FeatureFlags.MachineApi))
